func-apps/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# azure-functions
psycopg2-binary
openai==0.28.0
pydantic==2.10.6
pydantic-core==2.27.2